backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.12.0
anthropic==0.7.7
google-generativeai>=0.3.1
pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
httpx>=0.25.0

